extracted printing and exporting
